requirements.txt
asgiref==3.3.4
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-allauth==0.44.0
django-extensions==3.1.3
django-filter==2.4.0
django-rest-auth==0.9.5
django-storages==1.11.1
djangorestframework==3.12.4
flake8==3.9.2
idna==2.10
Markdown==3.3.4
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
Werkzeug==2.0.1
